+++ src/main/kotlin/com/xfastgames/witness/feature/JasmineBushFeature.kt
@@ -1,10 +1,10 @@
 package com.xfastgames.witness.feature
 
-import com.xfastgames.witness.blocks.flowers.LilacBush
+import com.xfastgames.witness.blocks.flowers.JasmineBush
 import net.minecraft.world.biome.Biome
 import net.minecraft.world.biome.Biomes
 
-class LilacBushFeature : PatchOfBlocksFeature(LilacBush) {
+class JasmineBushFeature : PatchOfBlocksFeature(JasmineBush) {
 
     override val biomes: List<Biome> = listOf(
         Biomes.SWAMP,